Separated commands into separate files
